+++ formula-android/src/main/java/com/instacart/formula/fragment/FragmentFlowStore.kt
@@ -7,7 +7,7 @@ import com.instacart.formula.android.FragmentId
 import com.instacart.formula.rxjava3.RxStream
 import com.instacart.formula.android.Binding
 import com.instacart.formula.android.FeatureEvent
-import com.instacart.formula.integration.DisposableScope
+import com.instacart.formula.android.DisposableScope
 import com.instacart.formula.android.FragmentBindingBuilder
 import com.instacart.formula.android.FragmentEnvironment
 import com.instacart.formula.android.FragmentState